+++ mdformat/codepoints/_unicode_punctuation.py
@@ -808,7 +808,7 @@ UNICODE_PUNCTUATION = frozenset(
     )
 )
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # pragma: no cover
     import string
     import sys
     import unicodedata