+++ mdformat/renderer/typing.py
@@ -1,18 +1,15 @@
-from typing import TYPE_CHECKING, Any, Callable, Mapping, MutableMapping
+from typing import TYPE_CHECKING, Any, Callable
 
 if TYPE_CHECKING:
     from mdformat.renderer import RenderTreeNode
 
-
-RendererFunc = Callable[
-    [
-        "RenderTreeNode",
-        # There is a recursion here. This should be
-        # `Mapping[str, RendererFunc],` but mypy doesn't support this until
-        # https://github.com/python/mypy/issues/731 is implemented.
-        Mapping[str, Callable[..., str]],
-        Mapping[str, Any],
-        MutableMapping,
-    ],
+# There should be `mdformat.renderer.RenderContext` instead of `Any`
+# here and in `Postprocess` below but that results in recursive typing
+# which mypy doesn't support until
+# https://github.com/python/mypy/issues/731 is implemented.
+Render = Callable[
+    ["RenderTreeNode", Any],
     str,
 ]
+
+Postprocess = Callable[[str, "RenderTreeNode", Any], str]